Functions and tests converted
