+++ cpl/CMakeLists.txt
@@ -4,8 +4,10 @@
 set(HEADERS
     ${CMAKE_CURRENT_SOURCE_DIR}/inc/container.h
     ${CMAKE_CURRENT_SOURCE_DIR}/inc/core.h
-    ${CMAKE_CURRENT_SOURCE_DIR}/inc/ecmath.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/inc/debug.h
     ${CMAKE_CURRENT_SOURCE_DIR}/inc/matrix.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/inc/ostream_overloads.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/inc/sort.h
 )
 
 set(SOURCES
