--- src/container/to_set.cpp
@@ -1,40 +1,0 @@
-/* MIT License
- *
- * Copyright (c) 2022 Brandon Pacewic
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-#include <set>
-#include <unordered_set>
-
-template <typename ForwardIterator,
-          typename BaseIteratorType =
-              typename std::iterator_traits<ForwardIterator>::value_type>
-[[nodiscard]] std::set<BaseIteratorType> to_set(ForwardIterator first,
-                                                ForwardIterator last) {
-    std::set<BaseIteratorType> set_obj(first, last);
-    return set_obj;
-}
-
-template <typename ForwardIterator,
-          typename BaseIteratorType =
-              typename std::iterator_traits<ForwardIterator>::value_type>
-[[nodiscard]] std::unordered_set<BaseIteratorType> to_unordered_set(
-    ForwardIterator first, ForwardIterator last) {
-    std::unordered_set<BaseIteratorType> set_obj(first, last);
-    return set_obj;
-}
