--- src/output_stream/output_container.cpp
@@ -1,50 +1,0 @@
-/* MIT License
- *
- * Copyright (c) 2022 Brandon Pacewic
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-#include <iostream>
-
-template <typename T_container>
-void output_container(const T_container& container, const bool space = true,
-                      const bool new_line = true) {
-    for (std::size_t i = 0; i < container.size(); ++i) {
-        std::cout << container[i];
-
-        if (space && i < container.size() - 1) {
-            std::cout << ' ';
-        }
-    }
-
-    std::cout << (new_line ? '\n' : ' ');
-}
-
-template <typename T_container>
-void output_reverse_container(const T_container& container,
-                              const bool space = true,
-                              const bool new_line = true) {
-    for (std::size_t i = container.size(); i > 0; --i) {
-        std::cout << container[i - 1];
-
-        if (space && i > 1) {
-            std::cout << ' ';
-        }
-    }
-
-    std::cout << (new_line ? '\n' : ' ');
-}
