--- src/scripts/README.md
@@ -1,6 +1,0 @@
-# Scripts
-----
-
-### [Test Samples](https://github.com/BrandonPacewic/TestSamples)
-
-### [Debug and Run](https://github.com/BrandonPacewic/DebugAndRun)
