--- src/tools/dbg.cpp
@@ -1,54 +1,0 @@
-/* MIT License
- *
- * Copyright (c) 2022 Brandon Pacewic
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-#include <iostream>
-#include <string>
-
-template <typename A, typename B>
-std::ostream& operator<<(std::ostream& os, const std::pair<A, B>& p) {
-    return os << '(' << p.first << ", " << p.second << ')';
-}
-
-template <typename T_container,
-          typename T = typename std::enable_if<
-              !std::is_same<T_container, std::string>::value,
-              typename T_container::value_type>::type>
-std::ostream& operator<<(std::ostream& os, const T_container& container) {
-    os << '{';
-    std::string separator;
-
-    for (const T& item : container) {
-        os << separator << item, separator = ", ";
-    }
-
-    return os << '}';
-}
-
-#ifdef DBG_MODE
-void dbg_out() { std::cerr << std::endl; }
-template <typename Head, typename... Tail>
-void dbg_out(Head A, Tail... B) {
-    std::cerr << ' ' << A;
-    dbg_out(B...);
-}
-#define test(...) std::cerr << "[" << #__VA_ARGS__ << "]:", dbg_out(__VA_ARGS__)
-#else
-#define test(...)
-#endif
